feat: restrict user access to other users

+++ src/api/user/user.controller.ts
@@ -1,21 +1,23 @@
 import {
-  Controller,
-  Get,
-  Post,
   Body,
-  Patch,
+  Controller,
   Delete,
+  Get,
   Param,
   ParseIntPipe,
+  Patch,
+  Post,
+  UnauthorizedException,
 } from '@nestjs/common';
 import { ApiBearerAuth, ApiTags } from '@nestjs/swagger';
 
 import { Authorization } from '@/core/compositions/authorization.decorator';
 
-import { UserService } from './user.service';
-import { type UserDto } from './dto/user.dto';
+import { AuthUser } from '@/core/decorators/param/user.decorator';
 import { type CreateUserDto } from './dto/create-user.dto';
 import { type UpdateUserDto } from './dto/update-user.dto';
+import { type UserDto } from './dto/user.dto';
+import { UserService } from './user.service';
 
 @ApiBearerAuth()
 @ApiTags('Users')
@@ -29,7 +31,7 @@ export class UserController {
    * @returns A promise that resolves to a UserDto.
    */
   @Post()
-  @Authorization('Create', 'User')
+  @Authorization('create', 'User')
   async create(@Body() createUserDto: CreateUserDto): Promise<UserDto> {
     return this.userService.create(createUserDto);
   }
@@ -40,8 +42,16 @@ export class UserController {
    * @returns A promise that resolves to a UserDto.
    */
   @Get(':id')
-  @Authorization('Read', 'User')
-  async findOne(@Param('id', ParseIntPipe) id: number): Promise<UserDto> {
+  @Authorization('read', 'User')
+  async findOne(
+    @Param('id', ParseIntPipe) id: number,
+    @AuthUser() user: UserDto,
+  ): Promise<UserDto> {
+    // Ensure the user is updating their own information.
+    if (this.userService.$can(id, user) === false) {
+      throw new UnauthorizedException();
+    }
+
     return this.userService.findOne(id);
   }
 
@@ -52,11 +62,17 @@ export class UserController {
    * @returns A promise that resolves to a UserDto.
    */
   @Patch(':id')
-  @Authorization('Update', 'User')
+  @Authorization('update', 'User')
   async update(
     @Param('id', ParseIntPipe) id: number,
     @Body() updateUserDto: UpdateUserDto,
+    @AuthUser() user: UserDto,
   ): Promise<UserDto> {
+    // Ensure the user is updating their own information.
+    if (this.userService.$can(id, user) === false) {
+      throw new UnauthorizedException();
+    }
+
     return this.userService.update(id, updateUserDto);
   }
 
@@ -66,8 +82,16 @@ export class UserController {
    * @returns A string indicating the success of the deletion.
    */
   @Delete(':id')
-  @Authorization('Delete', 'User')
-  async delete(@Param('id', ParseIntPipe) id: number): Promise<string> {
+  @Authorization('delete', 'User')
+  async delete(
+    @Param('id', ParseIntPipe) id: number,
+    @AuthUser() user: UserDto,
+  ): Promise<string> {
+    // Ensure the user is updating their own information.
+    if (this.userService.$can(id, user) === false) {
+      throw new UnauthorizedException();
+    }
+
     return this.userService.delete(id);
   }
 }
